Revert "onyx: Add syscall_ret hack for older Onyx versions"
This reverts commit 7d14c29360cf809a267eafb78f9af06fa2677cd2.
This hack breaks large return values which aren't negative.

+++ src/internal/syscall_ret.c
@@ -3,8 +3,7 @@
 
 long __syscall_ret(unsigned long r)
 {
-	/* TODO: We need this cast because the kernel doesn't cast return values properly, just yet... */
-	if ((int) r > -4096UL) {
+	if (r > -4096UL) {
 		errno = -r;
 		return -1;
 	}
